+++ app/Components/mainscreen.js
@@ -1,8 +1,7 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import React, { useEffect, useState } from 'react';
 import {
-    Alert, FlatList, Keyboard, KeyboardAvoidingView, Platform,
-    StyleSheet, Text, TextInput,
+    Alert, FlatList, Keyboard, KeyboardAvoidingView, Platform, StatusBar, StyleSheet, Text, TextInput,
     TouchableOpacity, View
 } from 'react-native';
 import colors from '../config/colors';
@@ -88,6 +87,9 @@ function Mainscreen(props) {
     }
     return (
         <View style={styles.container}>
+            <StatusBar
+                backgroundColor={colors.primaryBg}
+            />
             <View style={styles.tasksWrapper}>
                 <View style={styles.titleWrapper}>
                     <Text style={styles.sectionTitle}>
@@ -116,7 +118,8 @@ function Mainscreen(props) {
             <KeyboardAvoidingView style={styles.writeTaskWrapper}
                 behavior={Platform.OS === "ios" ? "padding" : "height"}>
                 <TextInput style={styles.taskInput}
-                    placeholder={"Write a task"} value={textInput}
+                    placeholder={"Write a task"}
+                    placeholderTextColor={colors.uiTextColor} value={textInput}
                     onChangeText={text => setTextInput(text)} />
                 <TouchableOpacity onPress={addTodo}>
                     <View style={styles.taskAddWrapper}>
@@ -134,7 +137,7 @@ const styles = StyleSheet.create({
         backgroundColor: colors.primaryBg,
     },
     tasksWrapper: {
-        paddingTop: 80,
+        paddingTop: 60,
         paddingHorizontal: 20,
         fontSize: 20,
     },
@@ -145,18 +148,23 @@ const styles = StyleSheet.create({
     },
     sectionTitle: {
         fontSize: 26,
+        color: colors.uiTextColor,
         fontWeight: "bold",
     },
     clearAllButton: {
         fontSize: 15,
         fontWeight: "700",
+        color: colors.uiTextColor,
     },
     uIDateAndTime: {
         fontSize: 14,
         paddingHorizontal: 5,
+        color: colors.uiTextColor,
+
     },
     items: {
         marginTop: 30,
+        borderRadius: 10,
     },
     writeTaskWrapper: {
         position: "absolute",
@@ -164,27 +172,29 @@ const styles = StyleSheet.create({
         width: "100%",
         flexDirection: "row",
         justifyContent: "space-around",
-        alignItems: "center"
+        alignItems: "center",
+        paddingHorizontal: 20
     },
     taskInput: {
-        paddingVertical: 15,
+        paddingVertical: 10,
         paddingHorizontal: 30,
         backgroundColor: "#fff",
         borderRadius: 60,
-        borderColor: "#C0C0C0",
-        borderWidth: 1,
         width: 250,
+        color: colors.uiTextColor,
+        backgroundColor: colors.taskInputBg,
     },
     taskAddWrapper: {
-        width: 60,
-        height: 60,
-        backgroundColor: "#fff",
+        width: 50,
+        height: 50,
+        backgroundColor: colors.taskAddBtnBg,
         borderRadius: 60,
         justifyContent: "center",
         alignItems: "center",
     },
     addText: {
         fontSize: 25,
+        color: colors.taskAddtextColor
     },
 });