+++ app/config/colors.js
@@ -1,7 +1,21 @@
 export default {
-    primaryBg: "#E8EAED",
-    itemBg: "#fff",
+    // App Universel color:
+    primaryBg: "#082032",
+    taskTitle: "#fff",
+    uiTextColor: "#fff",
+
+    // Main screen components color:
+    taskInputBg: "#464660",
+    taskAddBtnBg: "#464660",
+    taskAddtextColor: "#FA8072",
+
+    // Task component Color:
+    itemBg: "#464660",
     squareBg: "#01937C",
-    circularBg: "#F54748"
-    
+    circularBg: "#F54748",
+
+
+    taskTitle: "#fff",
+    taskCompleteTitle: "#C1AC95"
+
 };