+++ app/Changelog.md
@@ -13,3 +13,9 @@
 - Added Clear all the todos at one time.
 - Task lsit will store in the localstorage.
 - After reload previous task list will apear in the screen.
+
+> Changelog: v-0.7:
+
+- Entire app color changed.
+- Task input section resized.
+- colors are more organized now.